requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
datasets==2.13.2
dill==0.3.6
en-core-web-sm==3.7.1
filelock==3.12.2
Flask==2.2.5
frozenlist==1.3.3
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.4
MarkupSafe==2.1.3
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.10
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.2
pandas==1.3.5
pip==23.3.2
pkg_resources==0.0.0
preshed==3.0.9
pyarrow==12.0.1
pydantic==1.10.13
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rouge-score==0.1.2
safetensors==0.4.1
scikit-learn==1.0.2
scipy==1.7.3
setuptools==39.0.1
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.66.1
transformers==4.30.2
typer==0.9.0
typing_extensions==4.4.0
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.4
Werkzeug==2.2.3
wheel==0.42.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.15.0
